A package that holds only the colors, so another may hold different ones
OnymTokens, OnymAccent and Color.dynamic move out of OnymBrand.swift
into a new zero-dependency OnymDesignTokens package. OnymDesign depends
on it and re-exports it, so all 758 call sites across the eight UI
packages and the app compile untouched — this changes nothing that can
be seen.

What it buys is a seam. An adopter who wants Onym in their own colors
copies the package, keeps the module name, changes the values, and
repoints one dependency. Nothing else in the graph knows.

The completeness check is the compiler: a replacement that forgets a
token has no protocol to fail to conform to, only 758 call sites that
no longer resolve.

forSender stays behind in OnymDesign. Mapping a pubkey onto the palette
is policy, not a value, and adopters should inherit it rather than
reimplement it.
